Rename packages : pivotTable to pivottable
