+++ galite-core/src/main/kotlin/org/kopi/galite/visual/dsl/pivottable/PivotTableLocalizationWriter.kt
@@ -15,7 +15,7 @@
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
-package org.kopi.galite.visual.dsl.pivotTable
+package org.kopi.galite.visual.dsl.pivottable
 
 import org.jdom2.Element
 import org.kopi.galite.visual.dsl.common.Actor
@@ -57,6 +57,9 @@ class PivotTableLocalizationWriter : LocalizationWriter() {
     self.setAttribute("ident", ident)
     if (label != null) {
       self.setAttribute("label", label)
+    }
+    if (help != null) {
+      self.setAttribute("help", help)
     }
 
     peekNode(null).addContent(self)
