+++ galite-core/src/main/kotlin/org/kopi/galite/visual/pivottable/Constants.kt
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.pivotTable
+package org.kopi.galite.visual.pivottable
 
 import org.kopi.galite.visual.form.VConstants
 import org.kopi.galite.visual.Constants
