+++ galite-core/src/main/kotlin/org/kopi/galite/visual/pivottable/UPivotTable.kt
@@ -16,17 +16,13 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.pivotTable
+package org.kopi.galite.visual.pivottable
 
-import org.kopi.galite.visual.VRuntimeException
+import org.kopi.galite.visual.UWindow
 
-/**
- * This class represents exceptions occurring during execution process.
- *
- * @param message the associated message. It's optional
- */
-class VNoRowException(message: String? = null) : VRuntimeException(message) {
-  companion object {
-    private const val serialVersionUID = 0L
-  }
+interface UPivotTable : UWindow {
+  /**
+   * Builds the pivot table;
+   */
+  fun build()
 }
