+++ galite-core/src/main/kotlin/org/kopi/galite/visual/pivottable/VNoRowException.kt
@@ -16,20 +16,16 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.pivotTable
+package org.kopi.galite.visual.pivottable
 
-import javax.swing.tree.DefaultMutableTreeNode
+import org.kopi.galite.visual.VRuntimeException
 
-class VPivotTableRow(val data: Array<Any?>) : DefaultMutableTreeNode() {
-
-  /**
-   * Return the object at column
-   *
-   * @param        column                the index of the column
-   * @return        the object to be displayed
-   */
-  fun getValueAt(column: Int): Any? = data[column]
-
+/**
+ * This class represents exceptions occurring during execution process.
+ *
+ * @param message the associated message. It's optional
+ */
+class VNoRowException(message: String? = null) : VRuntimeException(message) {
   companion object {
     private const val serialVersionUID = 0L
   }
