+++ galite-core/src/main/kotlin/org/kopi/galite/visual/pivottable/VPivotTable.kt
@@ -16,7 +16,7 @@
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
  */
 
-package org.kopi.galite.visual.pivotTable
+package org.kopi.galite.visual.pivottable
 
 import java.io.File
 import java.net.MalformedURLException
