+++ galite-demo/galite-vaadin/src/main/kotlin/org/kopi/galite/demo/taxRule/TaxRuleP.kt
@@ -23,7 +23,7 @@ import org.jetbrains.exposed.sql.transactions.transaction
 import org.kopi.galite.demo.database.TaxRule
 import org.kopi.galite.visual.domain.INT
 import org.kopi.galite.visual.domain.STRING
-import org.kopi.galite.visual.dsl.pivotTable.PivotTable
+import org.kopi.galite.visual.dsl.pivottable.PivotTable
 
 /**
  * Tax Rules Report
